chore: api routes, move app routes
