+++ src/components/offer-review/offer-review-section.tsx
@@ -8,6 +8,7 @@ import OfferReviewList from './offer-review-list';
 import Loader from '../loader/loader';
 import { useAppSelector } from '../../hooks/use-app-selector';
 import { AuthorizationStatus } from '../../models/authorization-status';
+import { ApiRoute } from '../../configuration/api-route';
 
 type Props = {
   offerId: string;
@@ -19,7 +20,7 @@ export default function OfferReviewSection({ offerId }: Props): JSX.Element {
   const [comments, setComments] = useState<ReviewModel[] | null>();
 
   useAsyncEffect((signal) => handleRequest(
-    () => appApi.get<ReviewModel[]>(`comments/${offerId}`, { signal }),
+    () => appApi.get<ReviewModel[]>(`${ApiRoute.Comments}/${offerId}`, { signal }),
     setComments,
     { [404]: () => setComments(null) }
   ), [offerId]);
