+++ src/configuration/api-route.ts
@@ -1,0 +1,8 @@
+export enum ApiRoute {
+  Offers = 'offers',
+  Comments = 'comments',
+  Nearby = 'nearby',
+  Login = 'login',
+  Logout = 'logout',
+  Favorite = 'favorite'
+}
